requirements.txt
PyOpenGL
PyOpenGL_accelerate
numpy
pygame
opencv-python